+++ src/main/java/com/atlassian/db/replica/api/SqlCall.java
@@ -1,0 +1,7 @@
+package com.atlassian.db.replica.api;
+
+import java.sql.*;
+
+public interface SqlCall<T> {
+    T call() throws SQLException;
+}