+++ src/main/java/com/atlassian/db/replica/api/SqlOperation.java
@@ -1,7 +1,0 @@
-package com.atlassian.db.replica.api;
-
-import java.sql.*;
-
-public interface SqlOperation<T> {
-    T execute() throws SQLException;
-}