+++ src/main/java/com/atlassian/db/replica/spi/DualCall.java
@@ -1,0 +1,14 @@
+package com.atlassian.db.replica.spi;
+
+import com.atlassian.db.replica.api.SqlCall;
+
+import java.sql.SQLException;
+
+/**
+ * Splits main/replica calls.
+ */
+public interface DualCall {
+    <T> T callReplica(final SqlCall<T> call) throws SQLException;
+
+    <T> T callMain(final SqlCall<T> call) throws SQLException;
+}